systemd: remove os-release from RRECOMMENDS
If os-release is changes (e.g. by setting BUILD_ID to a build
unique value) having os-release in RRECOMMENDS causes bitbake to
execute a bunch of tasks. In particular, due to the
initramfs-ostree-torizon-image -> systemd -> os-release dependency
chain it causes the initramfs image to get regenerated on every
build, which causes it to have a new checksum and hence updates it
in the OSTree. Avoid rebuild of systemd when os-release changes by
by removing os-release from RRECOMMENDS.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>