aktualizr: use lockfiles when fetching aktualizr
From time to time we see fetcher failure when building multiple
images in parallel with a shared download directory:
  ERROR: aktualizr-1.0+gitAUTOINC+a94a15e31f-7 do_unpack: Fetcher
  failure: Fetch command export PSEUDO_DISABLED=1; export ...
  /workdir/oe/tmp/work/armv7at2hf-neon-lmp-linux-gnueabi/aktualizr/1.0+gitAUTOINC+a94a15e31f-7/git/
  failed with exit code 128, output:
  fatal: repository '/workdir/downloads/git2/github.com.advancedtelematic.aktualizr/' does not exist

This is probably due to a race condition in the git submodule
(gitsm) fetcher. Work around by using an extra lockfile in the
download directory for the aktualizr do_fetch task.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>