image_torizon_ostree.bbclass: use := instead of =
I found a 'basehash changed' issue when I run:
'''
bitbake torizon-core-docker torizon-core-balena
'''

it's caused by the way we set OSTREE_COMMIT_BODY, the content of it is
obtained from some git command output, so it can be changed during
bitbake parsing and running tasks stages, which will lead a
'basehash changed' issue.

Actually, for this OSTREE_COMMIT_BODY, we only need to know it's value
during parsing stage, so replacing '=' with ':=' can fix this problem.

Signed-off-by: Ming Liu <liu.ming50@gmail.com>
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>