initramfs-framework: add psplash support
Add psplash support to the initramfs-framework. This adds a new
initramfs module called psplash which starts psplash very early
in initramfs. The psplash process creates a FIFO in /run, which
will be moved to the real root. That allows to communicate to
the same psplash process from the root file system.

Kernel arguments are supported too, e.g. using splashangle=90
to rotate the splash screen or use nosplash to disable psplash.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>