image-common: drop file installations
These files should be handled in package recipes instead of being
installed directly in image recipe. Or else there are many drawbacks
for doing this. For instance, it will miss the FILE dependency checking
among packages, and it will miss the systemd procedures like
'systemctl install/enable' which are in do_package task.

Signed-off-by: Ming Liu <liu.ming50@gmail.com>