recipes-devtools: add correct idna version to fix docker-compose
docker-compose requires version 2.20.x of the requests module that, in
turn, requires idna v.2.5.
In meta-openembedded/meta-python idna revision has been bumped to 2.8
but this prevents docker-compose from running on our images.
Add old release to meta-toradex-torizon and set 2.5 as preferred version
in docker compose recipe to fix the issue.

Signed-off-by: Valter Minute <valter.minute@toradex.com>