python3-docker-compose: drop PREFERRED_VERSION from recipe
The PREFERRED_VERSION should be set in a conf file rather than in
a recipe. The setting of PREFERRED_VERSION in python3-docker-compose
just would not take any effect, we can safely drop it since the
python-idna/python3-idna in meta-toradex-torizon would be picked up
any way due to the higher BBFILE_PRIORITY setting.

Signed-off-by: Ming Liu <liu.ming50@gmail.com>