torizon: make use of Toradex BSP distro
This makes use of the Toradex BSP distro layer. It gets rid of some
duplicated code (e.g. arm-defaults.inc) and makes sure the BSP distro
and TorizonCore stays aligned.

With this change, TorizonCore now is ultimately based of Poky, which
pulls in some new distro features. This commit makes sure that distro
features stay largely the same as before. The only differences are
that we now have largefile and bluez5 in distro features. The largefile
feature is a sensible distro feature for TorizonCore. The bluez5
feature has no effect on master.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>