apalis-imx6/colibri-imx6/imx6ull: drop machines from TorizonCore layer
Drop those machines from the TorizonCore layer. Now that we have
meta-toradex-nxp in our layer stack we can reuse the machines from
there.

Add use-mainline-bsp to make sure we get a mainline compatible
selection of device trees to build.

The colibri-imx7 machine in our layer supports eMMC and raw NAND
which does not allow to reuse meta-toradex-nxp directly. So this
machine is handled in a separate commit.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>