docker-app: add recipe
Support standalone docker-app, docker-app needs a cnab-app-base to be
ready on docker hub, but unfornately it only supports amd64 ARCH, to be
able to support ARM CNAB images, we introduce a compiling option
CNABIMAGE, that the users could choose to set up their own images.

For a Toradex user, after you bitbake docker recipe, a cnab-run binary
would be deployed to ${DEPLOY_DIR_IMAGE}, then you need deploy that
file to:

https://gitlab.int.toradex.com/rd/torizon-core-containers/cnab-app-base

and build and push a CNAB image from there.

Signed-off-by: Ming Liu <ming.liu@toradex.com>