docker-compose: ignore docker-compose pull failures
If docker-compose pull fails docker-compose up should still be run. Use
the - to allow ExecStartPre to fail.

Note: We probably want aktualizr to take care of when to pull containers
actually (and aktualizr in turn should be controlled by the actual
application). If fixed tags/sha hashes are used, then it should actually
not matter whether docker-compose pull is called or not. Leaving the
pull command there allows to preprovision a device without installing
any containers, and have the system downloading them on first boot.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>