torizon-common: set distro version according to new version scheme
Use DISTRO_VERSION to implement the new Toradex versioning scheme.
This follows SemVer to some degree. The pre-release is generated
depending on current build type (typically supplied by CI/CD). If
no build type is specified a developer build is assumed. In this
case a build number of 0 is used and a full DATETIME timestamp to
identify the image.

Note that this relies that the bitbake task execution is started
on the same day, otherwise we end up with different date timestamps!
To allow fixing this we set TDX_PRERELEASE weakly.

Using DISTRO_VERSION makes sure that the exact version number lands
in /etc/os-release as well as /etc/issue. However, on a developer
machine those timestamps might be outdated since we do ignore
changes in TDX_PRERELEASE (vardepsexclude). In our CI/CD environment
this is not an issue as a new TDX_BUILDNBR makes sure that
DISTRO_VERSION gets evaluated again, leading to new os-release and
issue files.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>