image_type_torizon: use distro version number in OSTree commit
Instead of using the image name (IMAGE_NAME) use the distro version
number (DISTRO_VERSION) in the OSTree commit subject. The commit
subject is displayed in the image list on the Toradex OTA update
system.

The machine name and image name is given by the OSTree branch and
hence does not need to be part of the commit subject.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>