image_type_torizon: deploy kernel/initramfs/devicetrees to /usr
Deploy the kernel, initramfs and devicetrees to /usr/lib/modules/$kver
as suggested by the current OSTree documentation. This also allows to
access the current kernel, initramfs and devicetrees from a booted
TorizonCore system easier (/usr/lib/modules is directly accessible).
The previous location was in /boot which wasn't straight forward to
locate on a booted system since the actual /boot in a booted OSTree
system is the deployed version of /boot.

Related-to: TOR-238
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>