u-boot-distro-boot: adopt to new devicetree location
Devicetrees are now deployed to a new location. This is mostly
transparent to the boot script as fdt_path is set by OSTree to reflect
this new location. However, with the move OSTree also dropped the
"devicetree-" prefix (the new location has this information already
in the directory name "dtb"). Drop the "devicetree-" prefix and make
sure to put the complete location in quotes.

Related-to: TOR-238
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>