torizon-users: add recipe
Add a new recipe to contain all torizon users, currently only one user
'torizon' is supported, drop the EXTRA_USERS_PARAMS from image recipes.

A major reason for doing this is that docker-app needs torizon user to
be present in sysroot so it could install its docker CLI plugin to the
user's home directory and run chmod against it, otherwise, the chmod
command will fail due to torizon user not existing.

Another benefit from this change could be that it provides a extension
that we could install other user specific files (like ssh-keys) from
this recipe.

Signed-off-by: Ming Liu <ming.liu@toradex.com>