linux-toradex-mainline: fix a wrong inter-task regression
A regression was introduced by commit b11d276a3:
[ linux-toradex-mainline: fix a wrong inter-task dependency ]

We dont have to delete the task do_kernel_checkout, that somehow causes
it could not be re-added in, this probably is a bitbake issue. But as
a workaround, we could drop the bb.build.deltask line, just calling
bb.build.addtask with the correct inter-task dependency would be fine.

Related-to: TOR-694
Signed-off-by: Ming Liu <ming.liu@toradex.com>