udev-toradex-rules: fix toradex-net-rename.sh bug
Interpret serial as base-10 to fix "value too great for base" errors
when serial starts with '0'. Shell interprets numerical constants
starting with 0 as octals and if there is a number outside of base-8 in
the serial, it throws an error.

We could use "10#" to treat the product ID as a decimal value but that
does not work with the BusyBox shell. Use expr $var + 0 which works
well with bash and BusyBox to convert to a decimal value.

A simple test to reproduce this error:
$ echo $((08+1))
bash: 08: value too great for base (error token is "08")

Related-to: TOR-710

Signed-off-by: Abdur Rehman <abdur.rehman@toradex.com>


(cherry picked from commit e9239c25756812e2b5cd61005df00525d6611415)