ostree-kernel-initramfs: support fitimage
When fitimage is being enabled, the initramfs image and devicetrees
should not be installed or packaged. And the fitimage path is being
controlled by OSTREE_KERNEL, so we need make it to be a global
variable, same with OSTREE_DEPLOY_DEVICETREE define them in
torizon-machine-custom.bbclass.

The meta-updater layer will also need a similar change.

Signed-off-by: Ming Liu <ming.liu@toradex.com>