networkd-static-conf: add recipe
Introduce virtual/network-configuration, it's being used in
meta-updater layer to decide which network configuration package should
be installed, for instance, connman, networkmanager, networkd-dhcp-conf
and so on.

Since qemu machines are using tap devices to set up network, and the IP
addresses are hard-coded to host, client: 192.168.7.1, 192.168.7.2, we
introduce a networkd-static-conf package to manage the static IP inside
qemu machine.

With this change, the network works well on qemuarm64.

Signed-off-by: Ming Liu <ming.liu@toradex.com>