initramfs-framework: mount rootfs with OSTree modern layout
The OSTree modern layout does not create unnecessary directories in the
file system root (e.g. there is no /dev anymore). Use a custom rootfs
mount script which does not rely on /dev to be present.

This is a backport of the pending upstream patch "initramfs-framework:
check successful mount using mountpoint".

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>