aktualizr: make sure aktualizr can write U-Boot variables
The aktualizr boot retry support needs to be able to write U-Boot
variables using fw_setenv. However, because we store the environment in
eMMC boot partitions which are set to read-only by default, we have to
unlock the partition before we can write to it.

Note: Normally this is done using a bash alias. But since aktualizr does
not use the shell to invoke fw_setenv, this does not work in the
aktualizr case.

Signed-off-by: Naeem Khan <naeem.khan@toradex.com>
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>