docker-ce: depend on boot-complete.target
Make use of the boot-complete.target synchronization point. Order the
unit docker.service before this target to make its start necessary for a
boot to be considered successful.

This makes sure that aktualizr.service (which is ordered after
boot-complete.target) only gets started if the docker.service is
successful, and hence a boot is only marked successful if Docker starts.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>