aktualizr: sync service file from meta-updater
Resync aktualizr.service with meta-updater as of commit id 66c1ab96896a.
This adds network-online.target as a dependency. We have not noticed an
issue in this area, but it is better to stay in sync with upstream.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
(cherry picked from commit 66239209a5b78f12a58e7a199ac95bfaa122e559)