cni: workaround RM_WORK issue
There is a flaw in cni build system that it downloads and stores the
source files with read only mode, this leads to a problem when RM_WORK
is enabled that the source files could not be removed.

This issue has been observed with go-1.13/1.14 build, a workaround
could be we add 'w' mode to all source files.

Related-to: TOR-1138
Signed-off-by: Ming Liu <ming.liu@toradex.com>
(cherry picked from commit 36da30c3f4f1043e0fab47619ba46262013933cc)