u-boot-toradex: save env in persistent storage
To be able to support SOTA rollback on qemu machines, we need save env
in persistent storage rather than in mem which is the default way.

To achieve that, we enable CONFIG_ENV_IS_IN_FAT=y in fitfatenv.cfg,
also move other configs to it instead of patching uboot source.

Revert a patch in u-boot source to restore env address and size, and
generate a initial uboot.env image in DEPLOYDIR, this is needed or else
we might get a 'Bad CRC' issue when running fw_printenv/fw_setenv.

Related-to: TOR-955
Signed-off-by: Ming Liu <ming.liu@toradex.com>