torizon: do not add -rt to OSTree branch name
By default IMAGE_NAME does not contain the distro name. This leads to
the problem that an image built for the same machine once with the
torizon and once with the torizon-rt distro leads to the same image
files ond disk. When storing in a single directory, this leads to
conflict. Therefor, we append "-rt" to the images built with the
torizon-rt distribution.

However, -rt in IMAGE_BASENAME makes it into the OSTree branch name,
where we add the distro already. This leads to awkward looking branch
names:

  torizon:4/colibri-imx6/torizon-rt/torizon-core-docker-rt/nightly

Introduce a new variable IMAGE_BASENAME_SUFFIX. Redefine IMAGE_NAME to
include IMAGE_BASENAME_SUFFIX after regular IMAGE_BASENAME. This makes
the OSTree branch name looking more sane again:

  torizon:4/colibri-imx6/torizon-rt/torizon-core-docker/nightly

Related-to: TOR-1191
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>