torizon-conf: add containers tags as environment variables
According to the defined containers naming/tagging convention, new
published containers won't have the "latest" tag anymore. So it will
be mandatory for the user to pull a new container image in the
command line using a tag.

So let's introduce environment variables for each TorizonCore compatible
container with the value of the tag that should be used in the command
line to pull new container images.

Related-to: TOR-1295

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>