scripts:setup-environment-internal: add back DISTRO to BUILDDIR
To avoid DISTRO being redundant in BUILDDIR and TMPDIR, we dropped it
from BUILDDIR in commit XXXX:
[ scripts:setup-environment-internal: drop DISTRO from BUILDDIR ]

but that leads to a situation that the distros torizon/torizon-upstream
are sharing a same DEPLOY_DIR which causes 'allarch' packages conflict
among distros. This might not break autobuild since we are overridding
DEPLOY_DIR for autobuild, but it's not convenient for developers who
wanna run different builds on their local machines.

Let's add DISTRO back to BUILDDIR and drop it from TMPDIR.

Related-to: TOR-1434
Signed-off-by: Ming Liu <ming.liu@toradex.com>