polkit: avoid starting polkitd
We enabled polkit distro feature but only wanted its library but not
polkitd, to save some disk space.

Let polkit.service to start /bin/true instead of polkitd (which is not
present on rootfs), this ensures other systemd services that requiring
polkit.service would not fail due to dependency missing.

Related-to: TOR-1447
Signed-off-by: Ming Liu <ming.liu@toradex.com>