polkit: remove polkit from DISTRO_FEATURES
We added polkit beccause udisks2 needs it. However, udisks2 only needs
polkit as a library. If we add polkit to the distro features this will
enable build options in other packages like NetworkManager and Dbus
which then depend on polkitd which we don't add to TorizonCore because
of its size. Therefore, this commit removes polkit from DISTRO_FEATURES
and allows to build polkit without installing the polkit daemon.

Related-to: TOR-1455
Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>