docker-compose-up: enable restart policy
If you power off the device when docker-compose is pulling containers,
it could fail to start next time with the following message:

failed to register layer: error creating overlay mount to
/var/lib/docker/overlay2/.../merged: too many levels of symbolic links

If you run docker-compose again, it will restart pulling the
containers, and the problem is solved. So to workaround this issue,
let's define a restart policy for the docker-compose service. If the
service fails, it will try to run again after 10 seconds.

Related-to: TOR-583

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>