ostree: add fdtfile support in uEnv.txt
The default U-Boot environment variable for device trees is fdtfile.
Also, when TorizonCore Builder applies a device tree to a TorizonCore
image, the variable fdtfile is defined in uEnv.txt.

This patch adds support in ostree to this variable, to make sure its
rollback variant (fdtfile2) is created in an update, and used in a
rollback situation.

Related-to: TOR-1610

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>