u-boot-distro-boot: save default devicetree file to use in rollbacks
Devicetree files are defined by default via the fdtfile variable. To do
a rollback from an image with a custom device tree to an image using the
default devicetree (defined in U-Boot), we need a way to recover the
fdtfile variable from the U-Boot environment.

That could be achieved by resetting the environment (env default
fdtfile), but it doesn't work on modules where fdtfile is dynamically
configured via CONFIG_PREBOOT (on Verdin for example).

So let's save the default value of this variable as fdtfile2 at boot
time, before loading uEnv.txt, so we can use it in a rollback situation.

Related-to: TOR-1610

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>