meta: introduce image VARIANT
We want to support a image VARIANT to /etc/os-release to indicate what
image it is running for users to check.

To achieve that, we add a 'VARIANT' variable in os-release recipe so it
will be present in /etc/os-release in rootfs, and we need substitute it
in tweak_os_release_variant function according to 'IMAGE_VARIANT's value
defined in image recipes.

Related-to: TOR-1687

Signed-off-by: Ming Liu <ming.liu@toradex.com>