aktualizr: introduce rollback-manager service
Rollback manager is a small service implemented in shell script and
responsible for managing rollbacks in TorizonCore.

Until now, the only adopted strategy to rollback from a bad update was
checking the Docker daemon via a FailureAction=reboot entry in the
Docker systemd service.

This new rollback-manager service introduces a more flexible way of
managing and communicating rollbacks to the user:

1. It will run after boot-complete.target and before aktualizr.service.
2. It will rollback if the boot-complete.target doesn't start
   sucessfully, so if Docker or other essential service like dbus fails,
   it will rollback.
2. It will make sure to clean-up U-Boot variables when an offline update
   is done (e.g. via TorizonCore Builder) and Aktualizr is not running.
3. It will generate logs and communicate to the user about rollback
   events.

This implementation also fixes two bugs:

1. TorizonCore was rebooting indefinitely if the Docker service fails and
   there isn't an update in progress.
2. TorizonCore was automatically rolling back an offline update (e.g.
   via TorizonCore Builder) when Aktualizr was not configured/running,
   because upgrade_available was not being cleared after the update.

Runtime tested on Apalis iMX6 and Apalis iMX8.

Related-to: TOR-1716
Related-to: TOR-1720

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>