ostree-kernel-initramfs: add ostree-devicetree-overlays package
Split devicetree overlays and overlays.txt to a standalone package
ostree-devicetree-overlays, this is needed because we only install
overlays.txt to ostree rootfs when fitImage is enabled, in which case,
ostree-devicetrees would not be installed, so we need a extra package
to contain overlays.txt.

Also drop reference to device-tree-overlay-filter, it has been dropped
from Toradex BSP meta layer.

Related-to: TOR-1700

Signed-off-by: Ming Liu <ming.liu@toradex.com>