u-boot-distro-boot: merge boot.cmd.in
Merge boot.cmd.in of qemuarm64 to other machines.

The reason for having two boot.cmd.in files was because qemuarm64 uboot
does not have a loadaddr env, so it has to use scriptaddr to load
/boot/loader/uEnv.txt.

Fix it by checking loadaddr present firstly, not not defined, then go
with scriptaddr.

Related-to: TOR-1750

Signed-off-by: Ming Liu <ming.liu@toradex.com>