meta: fix issue when updating via Torizon OTA
This commit is basically reverting commit cea8d2177bc9.

Before the mentioned commit, we had the kernel parameters
"root=LABEL=otaroot rootfstype=ext4" in the boot script (uEnv.txt), and
with the mentioned commit we are now merging kernel arguments in
OSTREE_KERNEL_ARGS.

But this change had broke TorizonCore updatability from versions prior to
that change to newer versions (e.g. updating from last monthly release
to latest nightly release).

That is because, when a new OSTree commit is deployed (e.g. via
Aktualizr), it will use the kernel arguments from the field 'options' in
/boot/loader/entries/ostree-1-torizon.conf, and there we don't have
those parameters on images generated before the mentioned commit. Then
the kernel boots without those parameters, and the rootfs is not
mounted.

Related-to: TOR-1781

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>