meta: introduce fluent-bit agent
Fluent Bit is an open source Log Processor and Forwarder which allows
you to collect any data like metrics and logs from different sources,
enrich them with filters and send them to multiple destinations.

It will be used in TorizonCore core as part of Device Monitoring and
Fleet Management solution.

The systemd service is configured as disabled by default, and must be
enabled by the user or via a provisioning mechanism.

Related-to: TOR-1768

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>