greenboot: introduce framework to manage updates and rollbacks
This commit introduces greenboot (Fedora IoT project) as the framework
to manage updates and rollbacks in TorizonCore, deprecating the previous
rollback-manager service.

https://github.com/fedora-iot/greenboot

The implementation was done in a way that the behavior is the same as
before, so all the logic from the rollback-manager service was moved to
scripts called by the greenboot framework.

For the update to be successful, the systemd target boot-complete.target
must succeed. Additionally, the user may add scripts in
/etc/greenboot/check/{required.d,wanted.d} for additional boot checks,
scripts in /etc/greenboot/green.d to be executed when the boot checks
succeed, and scripts in /etc/greenboot/red.d to be executed when the
boot checks fail.

Related-to: TOR-1770

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>