u-boot-distro-boot: fix issue when updating verdin-imx8mm
On Verdin, the logic to define which device tree file to use is in
U-Boot. And sometime after the TorizonCore 5.1.0 release, the BSP
changed this logic (v1.1 was removed from the name).

So now, when you try to update TorizonCore from 5.1.0 to a later version
that has this change, it fails to boot due to a missing device tree.
Since U-Boot and its environment is not updated, the logic in U-Boot
will set a device tree that doesn't exist anymore in the latest images.

To fix that, let's rename the device trees when we identify that it is
using the old name.

Related-to: TOR-1935

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>