image_type_torizon.bbclass: drop prepare_ostree_rootfs prefunc
The purpose of prepare_ostree_rootfs prefunc is to ensure
do_image_rootfs runs from a clean ${OSTREE_ROOTFS} and call
oe.path.copyhardlinktree from ${IMAGE_ROOTFS} to ${OSTREE_ROOTFS}.

We dont have to maintain a prefunc to achieve that, it could be easily
done in do_image_ostree itself.

But the major reason for this change is to avoid hard links from
${IMAGE_ROOTFS} to ${OSTREE_ROOTFS}, which might lead to a Pseudo
abortion in some cases, which was observed when IMA/EVM is enabled
in rootfs.

There was a commit aiming to fix the Pseudo abortion, commit db09905:
[ image_types_ostree.bbclass: add fakeroot varflag to prepare_ostree_rootfs ]

but it's incomplete per later tests, there is still the case some old
inodes being used in ${IMAGE_ROOTFS} when do_rootfs runs, which causes
'rm -rf ${IMAGE_ROOTFS}' fail on a Pseudo abortion error.

This fixes the Pseudo abortion completely.

Reference:
https://github.com/advancedtelematic/meta-updater/pull/810

Related-to: TOR-2105

Signed-off-by: Ming Liu <ming.liu@toradex.com>