docker: build from the moby project
The Docker CE repository (https://github.com/docker/docker-ce) is
deprecated and not maintained anymore, so let's build Docker from the
Moby project.

For this, we are leveraging the recipe from meta-lmp, and applying some
changes on top of it.

With this change, the Docker version is being bumped to 20.10.5.

Related-to: TOR-2195

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>