networkmanager: fix build with enabled ppp
If 'ppp' packageconfig option is enabled, but the build system does NOT
have pppd binary installed, the build fails with:

| Has header "pppd/pppd.h" : YES
| Program pppd /sbin/pppd /usr/sbin/pppd found: NO
|
| ../NetworkManager-1.36.2/meson.build:570:4: ERROR: Assert failed: pppd
required but not found, please provide a valid pppd path or use
-Dppp=false to disable it

This is due to meson trying to look for the 'pppd' binary in the build
system when it should not. If the build system does not contain pppd,
the build fails.

This commit was backported from meta-openembedded/kirkstone-next:

https://git.openembedded.org/meta-openembedded/commit/?h=kirkstone-next&id=7040cffbb06b

Related-to: TOR-2323

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>