setup-environment: fix variable duplication on subsequent sources
There was a specific scenario where if you had already sourced
setup-environment before (whenever it was) and source it again now,
conf/auto.conf could end up with some duplicated variable definitions.

Since those are always appended to the end of the file it seems no issue
would be caused, but this could potentially mean that the file would
grow without end.

Related-to: TOR-2403

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>