setup-environment: fix EULA processing
The EULA for NXP has changed directory and the old rule was failing to
find it. Now, we present it to the user to read and accept it while
sourcing setup-environment, as well as an environment variable that
they can set and accept it without opening the EULA content.

Related-to: TOR-2402

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>


(cherry picked from commit a7622e6fc43daf3d5b79184b11bfc0c9818a9739)