python3-dotenv: convert recipe to bbappend
The python3-dotenv recipe is very similar to what is provided in
meta-virtualization, but since we are building a more recent version, we
cannot drop ours to use upstream.

But we can convert it to a bbapend file to make maintenance easier.

Related-to: TOR-2214

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>


(cherry picked from commit 5c8d82224892737e8223f370d7a6ecd5dddb7e07)