python3-docker: convert recipe to bbappend
The python3-docker recipe is very similar to what is provided in
meta-virtualization, but since we are building a more recent version and
applying a patch, we cannot drop ours to use upstream.

But we can convert it to a bbapend file to make maintenance easier.

Related-to: TOR-2214

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>


(cherry picked from commit e073cfc07f1de48be6199c3c764c7d05f5d60801)