setup-environment-internal: set use-head-next at distro level
When setting use-head-next in MACHINEOVERRIDES, only target builds will
get this override.

This is a problem because we do native builds of aktualizr-torizon that
also uses use-head-next to build from latest. And when SRCREV is
outdated, target builds will build from latest and native builds will
build an outdated version.

To fix that, let's set use-head-next at a distro level, so both target
and native builds will get use-head-next in OVERRIDES.

Related-to: TOR-2434

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>