linux-yocto: enable additional kernel options
Ideally, we would use our kernel cache for this. But linux-yocto kernel
has its own kernel cache repository, and currently, it's not possible to
have two kernel cache repositories and merge them.

Since we want to have a kernel for QEMU with a similar configuration
used in other machines, let's have a custom description file (.scc) to
enable additional kernel options not enabled in the default kernel
configuration provided by OE.

For now, this will only be used on QEMU builds and might be useful in
the future when integrating new machines that also use a kernel cache
repository.

For Toradex machines, this doesn't apply and the toradex-kernel-cache is
still the source of truth for kernel configuration.

Related-to: TOR-2455

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>


(cherry picked from commit ac6cdfc2d48f4b6242c41d931f34a1740ad82c9e)