packagegroup-core-full-cmdline: remove gawk and mc
gawk and mc are not useful in TorizonCore and can be safely removed.

This also removes perl from TorizonCore (perl was being included in the
image due to the dependency on mc-helpers-perl) and it saves around 15M
in storage.

Related-to: TOR-2325

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>


(cherry picked from commit 417dece6886f27eff63dbcf80ebee15f3e093b2c)