torizon-core-docker: move to Docker Compose V2
Docker Compose V1 is marked as deprecated and it's currently only
receiving patches for high-severity vulnerabilities or critical bugs.
Its support will be dropped soon by the community.

Docker Compose V2 is the currently active version maintained by the
community. After some investigation and intensive tests, we concluded
that V2 is backward compatible with V1 for TorizonCore use cases. So we
can safely move to it.

Related-to: TOR-2639

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>