u-boot: Generate unsorted environment file
Generate an unsorted environment file to allow for the search of a
binary blob inside the u-boot binary. This is done as part of the
bootloader update feature which will require that the environment
be reset before the switch to TorizonCore 6 (OS major upgrade).

Related-to: TOR-2592

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>


(cherry picked from commit e383f17f8a6958b26597e6819a8cb55a359d8793)