u-boot: Generate unsorted environment file
Generate an unsorted environment file to allow for the search of a
binary blob inside the u-boot binary. This is done as part of the
bootloader update feature which will require that the environment
be reset before the switch to TorizonCore 6 (OS major upgrade).
This is a backport from the work done for TOR-2592 on the kirkstone
branch.

Related-to: TOR-2656

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>