image_type_torizon: Add environment location and size information
Add to u-boot-ota.json information about location and size of the
u-boot environment inside the u-boot binary. With this information one
can extract the default environment directly from the binary avoiding
the need for a separate file with that data. This is a backport from
the work done for TOR-2592 on the kirkstone branch.

Related-to: TOR-2656

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>