lshw: cherry-pick patch to fix 'std::length_error' exception
lshw is failing on most modules with the following error:

terminate called after throwing an instance of 'std::length_error'
  what():  basic_string::_M_create
Aborted

This is because the get_string() function is wrongly using an unsigned
variable to check the return of the read() function, that might return a
negative value on error.

The fix is already in upstream:

https://ezix.org/src/pkg/lshw/commit/37c3f3934b9d0

Cherry-pick it to our layer while a new lshw version is not released.

Related-to: TOR-2665

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>


(cherry picked from commit 82db4ec4498fb7a43b4c81dbf0182602f70efdda)