systemd: use /dev/watchdog1 on Colibri iMX7
We have the systemd watchdog feature enabled in TorizonCore, and systemd
uses by default /dev/watchdog0 as the watchdog device.

Colibri iMX7 has two watchdog devices: one from the SoC (/dev/watchdog1)
and another from the PMIC (/dev/watchdog0).

Using the watchdog from the PMIC is causing problems when trying to
shutdown the device (shutdown -h now). The shutdown command is resseting
the device instead of shutting it down.

So let's switch to the internal watchdog device from the SoC.

Related-to: TOR-2650

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>