image_type_torizon: Generate link to bootloader binary for OTA use
Generate links for bootloader binary and corresponding version file
with fixed names (i.e. names that do not depend on the machine); with
this the image deployment directory will contain files 'u-boot-ota.bin'
and 'u-boot-ota.json' (symbolic links) - the former is always the
binary suitable for OTA use. This is a backport from the work done for
TOR-2490 on the kirkstone branch.

Related-to: TOR-2656

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>