polkit: remove patch
The objective of the patch is to change polkit's buildsystem to not
build polkitd, since TorizonCore doesn't need it.

Maintaining this patch has been a burden, since changes in the upstream
project (most of the times) require a rebase. The situation is worse now
that the project moved to Meson, and the whole solution would need to be
re-implemented.

Anyway, for this specific case, we don't need a patch. We can just
customize the install task to not deploy the polkit daemon and related
files.

Related-to: TOR-2749

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>