Request immediate reboot during a bootloader update
Requesting an immediate reboot rather than scheduling a reboot for a
minute in the future (as before) helps prevent sporadic bootloader
update errors. This is because the delayed termination signal sent by
systemd might be delivered to the action handler script during a
subsequent complete-install event.

Additionally we are also ignoring a SIGTERM during the action handler
execution which helps improve the situation if the reboot is started
by some process other the action handler script itself.

Related-to: TOR-2754

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>