image_type_torizon: choose the highest TEZI_CONFIG_FORMAT value
This variable is being set by both BSP and TorizonCore layers. In the
past we needed a higher vlaue than BSP, so we hardcoded an attribution
to this variable.

Now that the BSP also needed to increment this value we got into
a situation where we need to choose the higher number.

Related-to: TOR-2782

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>