torizon-core-common: add iproute2 packages
In TorizonCore 5.x.y, the ip command is provided by iproute2 package,
which is installed via this chain of dependencies: NetworkManager ->
dhcp-client -> iproute2.

In TorizonCore 6.x.y, the ISC DHCP client was removed from oe-core and
NetworkManager switched to an internal DHCP client implementation. As a
consequence, the iproute2 package is not installed anymore, and the ip
command is provided by BusyBox.

This breaks compatibility, because a TorizonCore 5.x.y user might have a
script that only works with the ip command from iproute2 and might fail
when executed with the BusyBox version.

To fix that, install the iproute2 packages in the image, so the same ip
command is provided on both TorizonCore 5.x.y and 6.x.y.

Related-to: TOR-2923

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>