image_type_torizon: generate composefs image
When building with secure-boot support, generate composefs descriptor
image.

The image is stored in the OSTree deployment directory
(/sysroot/ostree/deploy/torizon/deploy/).

After the image is generated, its fsverity hash is calculated and added
to the kernel command line arguments, so the image can be validated at
boot time.

Related-to: TOR-2669

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>