sudo: change executable permission flags
By default, the sudo binary is installed with only executable permission
(111).

This is a problem for some filesystems that require read permission to
execute binaries (like composefs).

So let's change its permission to 755.

There are no side-effects or security implications with this change (755
is acuatlly the default permission for the sudo binary on most Linux
distributions).

Related-to: TOR-2669

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>