uEnv.txt.in: support debugging the bootscript
It's a common need to debug the bootscript (i.e drop to the U-Boot
prompt after loading the bootscript but before booting the Linux
kernel).

To make that easier, let's define a U-Boot variable called
bootscript_debug. With this, users can stop at the U-Boot prompt and run
the following command to debug the boot script:

> setenv bootscript_debug 1; boot

Related-to: TOR-2956

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>