torizon-signed: check the signature of FIT images
The signature of FIT images can be enabled by inheriting the
torizon-signed.bbclass.

The name and location of the keys can be configured by setting
UBOOT_SIGN_KEYNAME and UBOOT_SIGN_KEYDIR variables.

For now, disable this feature for colibri-imx6ull-emmc and
colibri-imx7-emmc because a U-Boot DTB is not generated for these
machines (a device tree in U-Boot is required to save the public keys
used to check the signature).

Related-to: TOR-1704

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>