ostree: fix fdtfile patch
The patch was incorrect causing a "/boot" to be prepended to the value of the
rollback variable fdtfile2. This caused any rollbacks to fail since the boot script
would not be able to find the rollback device tree file.

This fixes the behavior so that the value of this variable is correct.

Signed-off-by: Jeremias Cordoba <jeremias.cordoba@toradex.com>