docker: run after usermount service
When docker data on external storage is enabled, the usermount service
is responsible for mounting the external storage partition.

Because of that, the docker service should wait for the usermount
service to finish its execution, or it might start using /var/lib/docker
before it is properly mounted.

Cherry-picked from kirkstone [9dee80d2048d]

Fixes: TOR-3116

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>