rust: bump rust to 1.70.0 and add missing python_pyo3.bbclass
Due to optee dependencies, we were facing some rust issue when building
natively. There wa an issue with python_pyo3.bbclass setting the wrong
target, that was solved and we did not backport.

Backport Rust 1.70.0 recipes and classes from openembedded-core's
master branch. This is being backported as is without any modifications.

Related-to: TOR-2913

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>