go: bumping go to 1.20.6 and bringing missing bbclasses
Docker image was failing to build due to missing some Go bbclasses. This
was causing some parsing errors during build time. So we bumped to the
newest Go version and backported the missing classes.

Backport Go 1.20.6 recipes and classes from openembedded-core's
master branch. This is being backported as is without any modifications.

Related-to: TOR-2913

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>