recipes-images: align the images between TorizonCore Docker and Lite
Aligned what utilities were being offered in TC Docker images with TC
Lite, so the user experience is the same in both. This led to the
removal of torizon-core-container.inc, since that's what as the source
of the differences between the two images, and none of its contents were
container specific. The contents were moved to torizon-core-common.inc,
so all images can be aligned.

Related-to: TOR-2770