u-boot-distro-boot: break down uEnv.txt transformations
Initially the uEnv.txt.in file was both transformed (into uEnv.txt)
and installed (into its installation directory) in the same OE task
(do_install) which made it harder for upper meta-layers to change this
behavior without duplicating code. To avoid this situation, the
transformations were moved into the "do_compile" task. Additionally,
a small change to uEnv.txt.in was made to allow the initialization via
OE variables of the FIT configurations to be forcefully used; this will
be useful to the meta-toradex-security layer.

Related-to: TOR-3246

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>