ostree: fix build issue of ostree patches being apart from its recipe
ostree .bbappend was under recipes-sota and its patches were under
recipes-extended. This was causing a build error where the recipe could
not find the patch to be applied. Now the .bbappend and the other
patches/files used were all moved to recipes-extended.

Also removed 0003-ostree-fetcher-curl-handle-non-404-errors-as-G_IO_ER.patch since
this was a backport patch, and ostree has just been bumped to version
2023.7.

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>